refactor txn test
